Move CPU context implementations to a specific module
