+++ kernel/src/arch/x86/signal.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MPL-2.0
 
-use ostd::cpu::{CpuException, CpuExceptionInfo, UserContext};
+use ostd::cpu::context::{CpuException, CpuExceptionInfo, UserContext};
 
 use crate::process::signal::{
     constants::*, sig_num::SigNum, signals::fault::FaultSignal, SignalContext,
