+++ kernel/src/process/clone.rs
@@ -2,7 +2,7 @@
 
 use core::{num::NonZeroU64, sync::atomic::Ordering};
 
-use ostd::{cpu::UserContext, sync::RwArc, task::Task, user::UserContextApi};
+use ostd::{cpu::context::UserContext, sync::RwArc, task::Task, user::UserContextApi};
 
 use super::{
     posix_thread::{AsPosixThread, PosixThreadBuilder, ThreadName},
