+++ kernel/src/process/signal/mod.rs
@@ -19,7 +19,7 @@ use align_ext::AlignExt;
 use c_types::{siginfo_t, ucontext_t};
 use constants::SIGSEGV;
 pub use events::{SigEvents, SigEventsFilter};
-use ostd::{cpu::UserContext, user::UserContextApi};
+use ostd::{cpu::context::UserContext, user::UserContextApi};
 pub use pause::{with_signal_blocked, Pause};
 pub use poll::{PollAdaptor, PollHandle, Pollable, Pollee, Poller};
 use sig_action::{SigAction, SigActionFlags, SigDefaultAction};
