+++ osdk/tests/examples_in_book/write_a_kernel_in_100_lines_templates/lib.rs
@@ -12,7 +12,7 @@ use alloc::sync::Arc;
 use alloc::vec;
 
 use ostd::arch::qemu::{exit_qemu, QemuExitCode};
-use ostd::cpu::UserContext;
+use ostd::cpu::context::UserContext;
 use ostd::mm::{
     CachePolicy, FallibleVmRead, FrameAllocOptions, PageFlags, PageProperty, Vaddr, VmIo, VmSpace,
     VmWriter, PAGE_SIZE,
