+++ ostd/src/arch/x86/cpu/mod.rs
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: MPL-2.0
+
+//! CPU context & state control and CPU local memory.
+
+pub mod context;
+pub mod local;
