+++ ostd/src/arch/x86/trap/mod.rs
@@ -28,7 +28,7 @@ use spin::Once;
 use super::ex_table::ExTable;
 use crate::{
     arch::irq::{disable_local, enable_local},
-    cpu::{CpuException, CpuExceptionInfo, PageFaultErrorCode},
+    cpu::context::{CpuException, CpuExceptionInfo, PageFaultErrorCode},
     cpu_local_cell, if_tdx_enabled,
     mm::{
         kspace::{KERNEL_PAGE_TABLE, LINEAR_MAPPING_BASE_VADDR, LINEAR_MAPPING_VADDR_RANGE},
