+++ ostd/src/user.rs
@@ -2,7 +2,7 @@
 
 //! User mode.
 
-use crate::{cpu::UserContext, trap::TrapFrame};
+use crate::{cpu::context::UserContext, trap::TrapFrame};
 
 /// Specific architectures need to implement this trait. This should only used in [`UserMode`]
 ///
